+++ contents/ui/components/ProviderCostSection.qml
@@ -90,6 +90,7 @@ ColumnLayout {
             level: 4
             type: Kirigami.Heading.Type.Primary
             Layout.fillWidth: true
+            elide: Text.ElideRight
         }
 
         Components.PlainComboBox {